requirements.txt
# Core dependencies
telethon==1.34.0
flask==3.0.0
flask-cors==4.0.0
numpy>=2.1.1
pandas>=2.2.2
python-dotenv==1.0.0

# Additional utilities
pytz==2023.3
requests==2.31.0